+++ tests/testthat/test-size.R
@@ -70,6 +70,18 @@ test_that("can subset data frame columns", {
   expect_equal(vec_slice(df, 1L)$y, vec_slice(df$y, 1L))
 })
 
+test_that("can subset empty data frames", {
+  df <- new_data_frame(n = 3L)
+  expect_equal(vec_size(vec_slice(df, integer())), 0)
+  expect_equal(vec_size(vec_slice(df, 1L)), 1)
+  expect_equal(vec_size(vec_slice(df, 1:3)), 3)
+
+  df$df <- df
+  expect_equal(vec_size(vec_slice(df, integer())), 0)
+  expect_equal(vec_size(vec_slice(df, 1L)), 1)
+  expect_equal(vec_size(vec_slice(df, 1:3)), 3)
+})
+
 test_that("can modify subset", {
   x0 <- NULL
   vec_slice(x0, 1L) <- 1